WIP: enable configuring VMCS execution controls
This requires careful bit manipulation: some special control registers
hold the authorized values (which can be either 0, 1 or both). From
that, we must verify that the user don't try to set a bit to a forbidden
value, set the reserved bits to their only possible value or lookup the
default one.