Enable saving host state
I'm not sure yet what we need to save and what we don't. For instance, I
get a #UD when reading the base of FS or GS, so we don't save them for
now.

This commit also include a refactor for easier `vmwrite`: different
values have different size, and we also want to be able to use types to
differentiate between fiels (e.g. Host vs Guest state). To that effect,
I added a trait for each size, now by implementing the appropriate trait
a field is associated with the correspondding `vmwrite` instruction
(taking the expected type as input).