Fixes VMX Host state
With this commit the host state is correctly restored after a VM Exit,
enabling inspection of guest launch failure.
For now the host resumes on a landing pad, and crashes at the end of the
landing pad function execution. Ideally, we want to set the return RIP
to the address just after the VMLAUNCH instruction.