Upgrade to bootloader v0.10.12
The bootlader version 0.9 crashed on real hardware (because of the
memory map fixed-size, our hardware has too much memory regions).
Hopefuly this version will work as expected.

The migration requires modification of the tooling, now instead of the
usual cargo commands (`run`, `build`, `test`) we use a new version
prefixed by 'k' (`krun`, `kbuild`, `ktest`). See the bootloader crate
examples for more details.

NOTE: one of the test (breakpoint) is failling, not sure if it is due to
the migration?