Fix failling test (breakpoint interrupt)
The test (and all interrupts) was failling because the new bootloader
does not initializes the SS segment to 0, which causes it to point to an
invalid segment and is no longer ignored. The fix is simply to ensure
that it gets initialized properly.

See https://github.com/rust-osdev/bootloader/issues/190