Setup VGA support
The current setup is supposed to work with VGA text mode, which is not
supported by UEFI and therefore not supported by the bootloader v0.10
(which is compatible with both BIOS and UEFI).
The next step is to port the VGA support to work with the frame buffer,
using a simple fixed-size font to draw pixels directly to the
framebuffer.