Add a FrameAllocator trait for VMX module
The VMX module now defines its own FrameAllocator trait, which is
implemented by the VirtualMemoryAreaAllocator. This is a much better
abstraction that allows us to allocate frames one at a time (and not a
whole virtual memory area).

This change was done in prevision of the work on EPT, which will require
allocating page tables one at a time.