Add support for Extended Page Tables (EPT)
Previously, the VMM was able to virtualize itself but both the guest and
host were using the same CR3 mappings. This commit enables extended page
tables, therefore adding an extra translation layer between the host and
the guest.

To that effect, we now create a set of EPTs that we populate with an
identity Guest-Physical to Host-Physical mapping from physical address 0
to the last valid physical memory frame. Because of the identity
mapping the observed behavior is the same as before, despite the new
inderection layer.

In addition, this commit introduces some infrastructure for inspecting
EPT violations. Later, we can extend that infrastructure to inspect
other kinds of exit qualifications.