Introduce Walker trait
A Walker is a struct capable of walking a kind of page table, and to
perform actions on them. Page tables can be either regular page tables
or extended page tables. Walkers can even be composed to simulate two
level page walk.
The two main methods, `walk` and `walk_range`, accept a callback that
will be called to each entry. This mechanism is flexible enough to cover
most use cases:
- Debugging by printing entries.
- Creating a new mapping by modifying the last level entry, and
  allocating intermediate pages if necessary.
- Selectively undoing mappings.

This commit also include an example of page table walker with a
`debug_print` method, to demonstrate walkers capabilities.