Demonstrate launch of a simple guest
This commit demonstrate the launch of a simple guest written in C.
Previously the vmm was only able to virtualize itself, therefore this
is a new milestone toward supporting fully fledged Linux guests.

The guest code is at the stage of proof of concept, and will be
refactored and refined as we improve our design abstractions.