Abstract frame allocator with a trait
This commit abstract away the SharedFrameAllocator, this will make it
much easier to change allocator in the future.
Our abstractions are kind of messy right now, we have multiple traits
for frame allocators each required by a different subsystem. We should
be able to clean this progressively as we develop our own walkers for
the various kinds of mappings.