Add RangeFrameAllocator
When allocating page tables for a guest, those pages must be allocated
from the guest memory. The current frame allocator does not give
flexibility regarding where the pages come from, and therefore is not
well suited for that task (currently we track all allocated pages and
add them to the guest EPTs).
This commit introduces a RangeFrameAllocator that can be constructed
from any physical memory range, and therefore can be told to allocate
frame from the guest physical memory.