VMX refactoring: lifetime & drop allocator dependency
A lifetime relationship between the different structures was added to
model the expected liveness of those structures. As example, the
VmcsRefion can no longer outlive the VMX mode (i.e. VmxOn must still be
alive).

Dependencies on the (vmx) frame allocator have also be removed for most
structs. Since we moved to a new EptMapper there is no need for a
specialized trait for frame allocation in vmx. Next commits will
completely remove the trait.