Initial support for Linux guests
This commit enable initial support for loading Linux guests. This
requires to load the elf segments at the required guest physical address
and to create the appropriate mappings so that they are accessible at
the required guest virtual address.

The current Linux guest causes a VM exit on CPUID, we will have to
emulate that instruction to move forward.