Handle Wrmsr for custom MSRs
Linux uses a few custom MSRs (such as 0x4b564d01) for KVM according to
the kernel doc. Those MSRs are outside of the range covered by the MSR
bitmasks and therefore always trap to the VMM.
For now those write are just ignored, but if we want to get KVM working
at some point we might need to emulate them properly.