Handle kernel command line
This requires inserting an additional payload into the guest (the
command line itself), which was impractical with the previous
implementation. In order to ease the process I refactored the guest
payload handling a bit, so that we can add an arbitrary number of
payloads in the guest using the guest memory allocator (which prevents
overlap).