Enabling fault injection
Tyche can now catch exceptions, print them, and reinject them into the
guest.
This is particularly useful for debugging linux boot (i.e., detecting
missmatches between cpuid and turned on features in the guest as was the
case with invpcid).
The guest trait now has a function to enable exception catching by
setting the exception bitmap.