Splitting VMM in two stages
For now the second stage is empty, and the first stage correspond to the
previous code (i.e. full VMM), but the structure is in place to start
operating the separation.
As part of the change, the tooling has been changed, we now use
justfiles instead of cargo to execute commands across the code base.