Refactor VMM to use identity mapping with guests
The memory is now divided in three portions:
- A lower portion, owned by the guest
- A middle portion, owned by the host
- A upper portion, owned by the guest

With this new strategy, the guest can access memory directly using an
identity mapping, which means it can directly access devices and read
hardware-provided structures, such as the ACPI tables, without need for
emulation or re-mapping.
The host (and sensible devices) is protected by not mapping its area
into the otherwise identity-mapped EPTs.

WARNING: for now the host is **not** properly protected! Indeed, the
Linux guest overwrite the VMM code in the current version. The reason
for this is that we don't really know where the VMM is located (we could
guess it's start and size maybe?), but that will be trivially fixed by
moving to the 2-stages model.