Jump into second stage
With this commit we jump into second stage instead of launching a guest
VM. The second stage will be responsible for launching the VM itself
once we move VMX there.

The transition to stage 2 is not complete yet for two reasons:

1. Stage 2 still uses the page tables of stage 1.
2. Stage 2 still uses the stack of stage 1.

We will need to resolve both of those points before we can safely launch
a guest VM from stage 2.