Define second stage ABI
This commit adds a new crate, stage_two_abi, that defines the ABI used
to bootstrap stage 2. That crate is used both by stage 1 and stage 2 so
that we can leverage the type system to provide a safe ABI between the
two. This crate is not intended to contain code in itself, and therefore
is not intended to grow bigger except for the Manifest definition.

The Manifest is a structure analog to boot_param for Linux, it defines
the configuration of stage 2 and is filled-up by stage 1. For now the
manifest is defined as a (hidden) static in stage 2, and is kept in the
elf despite being unused through some linking magic. Stage 1 can locate
the manifest by parsing the elf and fill the manifest before passing its
pointer to stage 2.