Refactor debugging tools
With the move to the 2 stages architecture a lot of the debug tooling
has to be adapted and/or can be simplified. For instance, we now used
identity mapping for the guest so we don't need all of the machinery we
had to use before to read guest physical or guest virtual addresses.