Cleanup some of the VMX initialization in stage 2
For our use case some of the initial values does not matter, for
instance the value of segments bases, limits and access rights can be
decided a compile time instead of reading the current values (in
particular, base and limit do not matter). Same for syscalls MSRs, the
guest will them up itself anyway.

Regarding host state, things are a bit more tricky, as we must not break
the environment. Some of the values are not usefull, like the syscalls
MSRs, but others like TR might be used? We could probably simplify
things even more using the fact that the IDT, GDT and TSS are static.