Setup infrastructure for GETSEC emulation
KVM does not support SMX extension, therefore we can't execute GETSEC
within the VM without causing a VM exit, in which case KVM will shut
down our VM.

To work around that and still develop within QEMU, we will partially
emulate GETSEC within stage 1 (and 2 if necessary). This can be done by
exploiting the fact that GETSEC causes a #UD exception if the SMXE bit
in CR4 is not set, hence the #UD handler can emulate the instruction
without relying on KVM emulation.

This commit does not provide emulation yet, but setup the infrastructure
to intercept GETSEC, read and modify registers and continue execution
past the GETSEC instruction.