Second-stage ABI refactoring
Refactor the guest main loop to achieve:
1. More architecture independent design
2. A cleaner way to handle VM Calls

The VM Call ABI is now defined in the `hypercalls` module, each
hypercall is identified by its own number and handled by the hypercall
module, independently of the guest backend.

With the current x86_64 ABI, the hypercall number is passed in %rax, and
the three parameters in %rcx, %rdx, and %rsi. The error code is returned
in %rax, while the up to three returns value are returned in %rcx, %rdx,
and %rsi.