Enable build of Risc-V binaries for stage two
This commit enables the build of binaries for stage two on the 64 bits
Risc-V target, but the binary itself can't be booted yet.

This is achieved mostly through refactoring and segregating
platform-dependant functionalities in two modules (`x86_64` and
`riscv`). In particular the common utilities of `vmx` have been moved
into a new `utils` crate that is common to both architectures, and most
of architecture specific functions of stage two have been moved in the
`x86_64` module.

The Risc-V binary can be build with `just build-riscv` and the target
was added to `just check`.