Adding a simple one step revocation mechanisms
This implementation is not meant to remain as is, but rather to allow
cleaning up an enclave to run multiple experiments without rebooting the
VM.

I added a revokInfo struct to the RegionCapa that tracks a domain and
handle pair.
If the capa is a revocation one, this points to the domain and handle
that can be revoked.
If the region is an access capa, it holds information about the domain
and revocation handle to reclaim access to the corresponding region.

NEXT: I need to implement support for revocation inside the enclave
driver, add a reclaim vmcall to tyche, and a merge function.