Prepare switch from PIC to APIC in stage 1
To support multiprocessor, we need to replace the legacy PIC with APIC.
This commit introduces a new dependency to initialize APIC directly.
In addition, we use an acpi crate to easily parce ACPI tables.

x2apic: https://github.com/kwzhao/x2apic-rs
acpi:   https://github.com/rust-osdev/acpi

To use the acpi library, we need to define a new type implementing
ACPIHandler. This commit implements the handler and defines the virtual
to physical address mapping function for the acpi library.

Use the acpi library and the ACPIHandler to parse the ACPI table.
Need more commits to replace the current usage of AcpiInfo with
AcpiTables.

Signed-off-by: YuChen Qian <yuchenqian96@gmail.com>