Split `allocate_range` out of FrameAllocator
Stage 2 allocator was not able to allocate ranges of physical memory
(and it doesn't need to), so the FrameAllocator trait made little sense
in the stage 2 context.
This commit split the FrameAllocator trait in two: a classic frame
allocator that can allocate frames one at a time, and a RangeAllocator
that can aditionnaly allocate ranges.
We should probably simplify the FrameAllocator even more in the future,
for instance remove the `get_boundaries` which doesn't make sense for
some kinds of allocators.