Free frames after unmapping range in EPT mapper
The unmap range function was correctly setting entries to 0, but pages
were not freed. This commit enable running a cleanup function on pages
that are entirely covered by the range in the new
`Walker::cleanup_range` method. To free frame, the cleanup simply has to
free the given frame. This will not free the root page, though.