acpi: place rsdp at boot parameter for guests
The guest (linux) failed to detect the existence of ACPI table and gives
the following error during the boot:
[    0.015192] ACPI BIOS Error (bug): A valid RSDP was not found (20210930/tbxfroot-210)

This creates several issues. Most importantly, Linux will try to boot
with the legacy PIC if it cannnot parse the APIC table inside ACPI.
As a direct consequence, the guest can never boot up with multiple
cores...

/ # cat /proc/interrupts
           CPU0
  0:      31353    XT-PIC      timer
  1:         10    XT-PIC      i8042
  2:          0    XT-PIC      cascade
  4:        245    XT-PIC      ttyS0
  8:          0    XT-PIC      rtc0
 10:          0    XT-PIC      i801_smbus
 12:        126    XT-PIC      i8042

This commit fixes the issue by placing the RSDP pointer to the boot
parameter, so the guest currently sees the same ACPI table read by the
host.

Now if we boot up linux with a single core, we can see it detects ACPI
table and switches to APIC. A quick verification on interrupt source:

/ # cat /proc/interrupts
           CPU0
  0:         44  IR-IO-APIC   2-edge      timer
  1:          9  IR-IO-APIC   1-edge      i8042
  4:        197  IR-IO-APIC   4-edge      ttyS0
  8:          1  IR-IO-APIC   8-edge      rtc0
  9:          0  IR-IO-APIC   9-fasteoi   acpi
 12:        125  IR-IO-APIC  12-edge      i8042
 16:          0  IR-IO-APIC  16-fasteoi   i801_smbus

Signed-off-by: YuChen Qian <yuchenqian96@gmail.com>