Modifies stage 2 entry point ABI
It is no longer needed to pass arguments to stage 2 (i.e. the manifest).
This will make it easier to jump into stage 2 from stage 1, and in the
future Risc-V and with trusted boot.
Stage 2 is now able to locate statics itself, and use helper functions
to protect from emiting multiple mutable references to the same static
value.