smp: load IDT for each AP
After loading the GDT, we can load IDT on each processor and specify the
interrupt handler for each occasion.

This commit splits the interrupts.rs into idt.rs and apic.rs (to be
committed soon). The idt.rs is solely responsible for the interrupt
descriptor table, and the apic.rs is responsible for the apic logic.

Signed-off-by: YuChen Qian <yuchenqian96@gmail.com>