Add support for VGA in stage 2
This commit enable booting stage 2 on bare metal with VGA support, by
reusing the driver extracted from stage 1.

NOTE: the transition was initially not working because of unexpected
values on some variables with initial value of 0. The cause was that the
BSS section was not properly initialized on bare metal (but on QEMU we
get pages that are zeroed out by the kernel, so we didn't run into the
issue earlier). To fix this, this commit simply zero-out the difference
between segments `filesz` and `memsz`, which by extension initializes
the BSS.