smp: disable 2nd SIPI
The 1st SIPI sends the AP to the trampoline and jumps to the entry
function to setup the CPU core. The 2nd SIPI causes the AP to enter the
trampoline again and the entry function will try to initialize the core
again. Such behaviour causes issue because the per-cpu structure Cpu is
only supposed to initialize exactly once...

This commit fixes the issue by only sending the 1st SIPI, as it is more
than enough to boot up the core under qemu.

Signed-off-by: YuChen Qian <yuchenqian96@gmail.com>