Refactor GDT and IDT handling in Stage 2
This commit is a work in progress toward #7. Previously, we has a single
GDT and IDT, both mostly empty, and there was no plan for multi-CPU
support. Now we still use a single IDT and GDT, but the GDT has enough
capacity to hold a descriptor for each of the different cores' TSS.
For now the TSS are not properly setup (they need to know about all the
cores' stacks, which will be passed in the manifest in the future), but
the descriptors are in place in the GDT.

What is left to do is:
- Fill each TSS with the stacks info
- Generate the TR selector for each core (the selector pointing to the
  core's TSS descriptor in the global GDT).
- Fill IDT with proper exception handlers.