smp: fix race condition in cpu initialization
The original check on whether the CPU is initialized or not is not
atomic, and causes race condition sometimes (probably due to context
switch after getting the cpu option value).

This commit fixes this issue by adding an array of atomics, and does an
atomic xchg to acquire the access into the index of the CPU array.

Signed-off-by: YuChen Qian <yuchenqian96@gmail.com>