stage-2: refactor guest and separate out vcpu
Separate out the vcpu related functions from the x86 guest of the second
stage, and create a new vcpu struct in a separate file. Makes it cleaner
and easier to manage when we have multiple cores.

Also refactored the hypercall structure to make sure it is only
instantiated once. Got rid of unnecessary boilerplate code to create
and get the global static mutable variable and both the region and
domain arena are initialized only once by the BSP.

Signed-off-by: YuChen Qian <yuchenqian96@gmail.com>