vcpu: enable unrestricted guest
The Linux trampoline code switches off the PG bit inside CR0 even if we
attempt to boot on the 64-bit trampoline directly. Some of the observed
bits from gdb:

- Initially, CR0 = 0x8000_0031 [PG NE ET PE] set by VMM
- CR0 = 0x0005_0033 [AM WP NE ET MP PE] at the initial stage of the
  Linux AP trampoline (turns off paging)
- CR0 = 0x8005_0033 [PG AM WP NE ET MP PE] at the late stage of the
  Linux AP trampoline (turns on paging)

This requires us to enable the unrestricted guest feature in VMX so that
the CR0.PE and CR0.PG can be controlled by the guest.