Makes physical memory offset globally available in stage 1
The physical memory offset (i.e. offset between virtual and physical
memory) in stage 1 is fixed by the bootloder and never updated, but it
is only available from the bootinfo and the allocate trait. This commit
makes it available from any ware in stage 1, provided that the memory
subsystem was initialized.