Automate debugging process through justfile
Due to recent updates to the justfiles, the gdb socket changed to a new
per-user scheme. This makes it possible to run Tyche on the same machine
for different users simultaneously, but also made debugging tricker as
the socket has to be explicitly passed. This is no longer necessary with
this change.

Example usage:

```sh
just linux-dbg

just gdb
```