acpi: add mp wakeup structure
Define the wakeup entry and the mailbox structure inside the first stage
acpi parsing utilties. To add the MADT wakeup entry inside the original
MADT, we first allocate a new page for the MADT (double the size), and
then copy the original over. After that, we add the 0x10 (wakeup) entry
at the end and change the length and checksum of the MADT.
Correspondingly, we also need to change the checksum of XSDT, the MADT
entry of RSDP and the checksum in RSDP.

Signed-off-by: YuChen Qian <yuchenqian96@gmail.com>