Restore AP trampoline frame content after CPU boot
The frame used for the AP trampoline needs to be in the low memory, but
this regions is also used by the bootloader to store potentially
important data (e.g. page tables).

To minimize the risk of crash, the frame used fot the AP trampoline is
now backed up and then restored once the boot process is complete. This
is definitely hacky, but the alternative is to either require help from
the bootloader (so we would need to fork it) or duplicating all required
structures (mainly PTs).