acpi: add mp mailbox wakeup logic under 2nd stage
Let the AP spin on the command field and the APIC ID field of the
mailbox page in root mode and wait until the guest OS issues a wakeup on
the current core. Once the core is ready to bootup, we read the wakeup
vector from the mailbox and change the RIP of the vCPU and launch it
under non-root mode.

Signed-off-by: YuChen Qian <yuchenqian96@gmail.com>