smp: map ap trampoline page to guest allocator
Originally we don't have the trampoline page identity-mapped to the
guest (Linux) page table, so it incurs a triple fault vmexit and gdb
cannot access the memory of the trampoline. It would be a better fix if
we check the spec closer so that we don't hard-code the trampoline page
address.

Signed-off-by: YuChen Qian <yuchenqian96@gmail.com>