ipi: add hypercall to test host<->host ipi
This commit modifies the libtyche to add a new vmcall instruction to
send ipi between hosts. In addition, we have added a ipi sending
function on the host side. The test hypercall directly triggers this ipi
sending function on the host CPU, waiting for other vCPU to trigger a
vmexit back to the host for the host-to-host ipi delivery.

Signed-off-by: YuChen Qian <yuchenqian96@gmail.com>