Simplifies VMX lifetimes
This commit removes the `'active` lifetime of ActiveVmcs by simply
moving the VmcsRegion in and out of an ActiveVmcs instead of using
references.
To further simplify manipulations of VMCS structures, an unsafe version
of `create_vm` is added to return a `'static` VmcsRegion. The catch is
that the compiler no longer ensures that VmcsRegion outlives Vmxon, so
this must be enforced manually when using this new function.