elf: round down the address
When testing on the raw hardware, one of the second stage ELF segement
is not aligned to the page size (0x0ff98) and causes pt_mapper to panic.
This commit fixes this issue by rounding down the starting address of
the segment and increase the number of bytes of the segment if
necessary.

Signed-off-by: YuChen Qian <yuchenqian96@gmail.com>