mmu: enable caching on ept entries
We can specify the EPT memory type to enable/disable caching on each of
the leaf entries. Previously, we disable the bit and linux was not able
to boot up (super slow)... To fix this issue, we specify a WB policy for
the EPT memory type on the leaf nodes.

Signed-off-by: YuChen Qian <yuchenqian96@gmail.com>