Stage 2 x86_64 vcpu refactoring
The vCPU where stored inside of the monitor state, which means that they
were behind a mutex and that only one cpu could run the guest at a time.
This commit extracts the vCPU from the state and instead uses one vCPU
per physical CPU.

Multi-core is currently triple-faulting when a CPU is woke up by the
guest.