Skeleton for the new version of the driver
This driver's goal is to manage both enclaves and capabilities.
It has a simpler api than the previous approach as the workflow is
supposed to be:

1. Create the enclave

This only creates internal data structures but does not call into the
capability and/or tyche.

2. Mmap a region for the enclave

This reserves contiguous physical memory pages and maps them into the
user process.
Still, it does not create any domain and/or perform capability
operations.

3. Get the enclave physical offset from the driver

4. User process populates the content of the enclave and uses calls
   similar to mprotect to create the various segments

The driver will accumulate the different segments and protection rights.

5. Commit

That's when we actually call into the capability library.