Create return switch handle on domain transition
The handle used for the transition is removed form the caller domain and
a new "return" handle is inserted in the callee domain.

The current version do not check that the target context is not already
running, this should be added in a future version. I'm not sure if it's
best to bake it into the capa engine or to keep it outside. If it's
added inside, then we can output precise update domain instructions
(like update access rights on core n°X).