Registering a handler for trap int3 (debug)
Apparently Linux is not willing to forward the segfault, so for the
moment we transform it into an int3 (hardware debug).

Ultimately, when have a story for the interrupts, sending a segfault for
an EPT violation seems okay (Linux will kill the offending process).
We might need to make a few adjustments for clean up, but other than
that it seems like the appropriate behaviour.
After all, it's a segfault.