Implementation of segment & syscalls
This C folder provides code snippets to manipulate the configuration of
ring 0 inside the domain.
This will probably not be the final version of the TRT, which we will
probably end up writting in Rust, but rather an exploratory
implementation so that we have reference code to manipulate all the
necessary registers.

Next: I will implement some user space transition probably.