Prepare context switch on interrupt
For now progress is blocked until we decide whether to stick to context
of move toward a one context per core model.
With the current context approach, the monitor is responsible for
multiplexing contexts on cores, which seems like a higher level function than
what we should implement in a security monitor. For instance, we need to
register a context for handling interrupts on each of the cores in order to
prevent re-entry.