Routing interrupts
I am getting closer to routing interrupts.

The problem for the moment is that we do not have the machiney in
user-space/sdkenclave to handle the exceptions.
The next commit should fix that.
I plan to extend the infrastructure to register handlers for such cases
by default, but also to have a selected execution of the enclave's code
and the outside so that we avoid duplicating too much code.