Refactoring following Charly's feedack.
This is a big commit with the following changes:

1. Make configuration more generic and handle all cores
  a. Seal no longer takes cr3, rip, rsp
  b. configuration calls are used to set perm, switch mode, traps, and
  cores.
  c. A new vmcall set_entry_on_core allows to set cr3, rip, rsp for
  a domain's context on a particular core.
2. I had to adapt the C libraries as well to match the new interface.
3. Contexts have an RCFrame (reference counted frame) to their vmcs and
   we support all three models (shared, copied, fresh).

TODO: for fresh vmcs, we should do a minimal initialization which isn't
done yet.

Next: I should also clean-up the C code to simplify it. We really don't
need all that code inside the driver and could get away with doing most
of it in user space.