kvm: fresh vcpu setup
In the case when we have a switch type of FreshVCPU, we first save the
frame address of the orignal frame, and then switch to the new vmcs
frame and setup the default vmcs values (taken from kvm-intel).
Afterwards, we switch the vcpu back again to the original vmcs frame.

I'm thinking about when to actually start these fresh vcpus...

Signed-off-by: YuChen Qian <yuchenqian96@gmail.com>