I/O: create I/O domain from hypercall
This is a shortcut to create an I/O domain from the dom0 guest through a
hypercall. It would probably be much better to have a global I/O domain
right inside tyche, but I have a bit trouble getting the domain handler
on the guest kernel side. As a result, I opt to add this commit for now
(and probably revert this one later).

Signed-off-by: YuChen Qian <yuchenqian96@gmail.com>