capa-engine: Allow in-place re-allocation of a domain
Our arena allocator requires to pass the new object to allocate to
ensure all invariants are respected. But because domains are huge struct
they can cause stack overflow.
This commit enable allocator without passing an explicit object, with
in-place reset of the arena slot. The caveat is that we need to manually
reset the fields with non-default values, in the case of domains the
domain ID.