Fixing interrupted resume
There was a bug.
An interrupted domain gets resumed as a regular synchronous transition,
meaning we override rdi to present the domain with a return handle,
which obviously disrupts the domain's execution when it actually resumes
as the register is now polluted.

To fix this, I mark a context as interrupted in handle_violation.
If marked, upon a resume, rdi is not modified (and the interrupted flag
for the context is cleared).