I/O: create I/O domain inside tyche
Since we disallow a guest to create an I/O domain by itself, this commit
directly create a single instance of I/O domain inside the microhypervisor.
As a result, only tyche keeps track of the IOMMU. Only when TD0 boots
up, we will have a new IOPT mapping sent to tyche.

Signed-off-by: YuChen Qian <yuchenqian96@gmail.com>