Capa-engine: use a global region pool instead of per-domain
Remove the per-domain region pool in favor of a global pool. This
drastically reduces the memory footprint and makes it possible to use
much more regions per domains.

A possible drawback is that the current implementation allows for DoS if
a domain consumes too much of the regions from the global pool. This is
mitigated by the maximum number of capabilities per domain, but we could
also set a counter to allow limiting the number of regions for a given
domain.

Potential regressions: used regions might not yet be properly cleaned
(it used to be automatic when creating a new domain), this will be part
of my next focus area while working on better support for
transactionally.