Initial alias and carve implementation
This patch introduces a new region capability which allows carving and
aliasing. It is the first change as part of the capa-v5 effort.

The new region is not tight to a binary tree like the previous version,
but can instead have an arbitrary number of children. A region can be
aliased, to create a new region with shared access, or carved to create
a new region with exclusive access.

This patch adds the new region type alongside the old one, but do not
replace it yet