Capa-engine: add capacity check to gen arena
It is now possible to check for available capacity in generational
arenas. This commit uses the capacity check to ensure the carve and
alias functions never fail due to out of memory errors on region capa or
local capa allocations (but there might be out of memory errors in the
region tacker!).
In the future, those check will be extended to all exposed interfaces.