Capa-engine: add remapper to decouple remapping from capa-engine
This commits adds a new Remapper struct which holds a list of remapping.
The remapper can be used to wrap a permission iterator (the one used to
updates EPTs and PMPs) and add a layer of indirection.
The remapper is notably **not** part of the capa engine, it has to be
instantiated separately. It is more like a thin layer on top which can
be used to facilitate emulation of hypervisor-like functionalities.