tyche: push TlbShootdown upon init signal vmexit
We have to push a TlbShootdown update to the core's update queue to
ensure to process a TlbShootdown at the current vmexit cycle.

Otherwise, we might end up in race conditions. For example:

1.
  - Core 0 push TlbShootdown { core = 2 } into the update queue
  <- Possible Vmexit on core 2 and processes the Tlbshootdown early,
     updating to the original ept root
  - Core 2 process the TlbShootdown update in apply_core_updates and
    update the ept root

2.
  - Core 2 process the InitSignal Vmexit, but Tlbshootdown is not yet in
    the update queue.

It would be nice if we can have a clear and moduar separation between
the capa engine and tyche. However, in this case, the code has to be a
bit mingled together...

Signed-off-by: YuChen Qian <yuchenqian96@gmail.com>