tyche: tlb flush barrier per domain
Introduces one TLB flush barrier per domain, and use a per-domain atomic
boolean to restrict to one TLB flush at a time. The rest of the core
trying to initiate a TLB flush has to spin on the atomic before doing
the IPI and process the vmexit.

The initiating core is now in charge of the whole TLB flush process,
which means that the core initiating the EPT root update will be
responsible for acquiring the EPT_FLUSH lock, pushing the TlbShootdown,
creating the barrier (one single barrier per update), sending out the
IPI, waiting on the barrier, freeing the original EPT root, and release
the EPT_FLUSH lock.

Note that we have to add an extra `init` to differentiate the EPT root
update during the initialization process. Tyche allocates and
initializes the EPT root first before initializing and adding the vCPU
into its `cores` domain bitmap. This is problematic because we don't
want to send out IPI before the vCPU are created (and Tyche updates the
EPT root when we create the vCPU at the startup, so we don't need a
flush)

Signed-off-by: YuChen Qian <yuchenqian96@gmail.com>