Fixing the interrupt bug
:warning: smp is broken...

I think the APIC passthrough might be confusing the host's KVM.

What happened:

Dom1 was booting all the way up until the first call to idle, which
triggered a htl.
The hlt gets trapped all the way into dom0 that calls vcpu_block.
The vcpu_block waits for a timer interrupt to trigger the list of
deadlines for registered timer event, including waking up dom1
(hrtimer).
The problem was that dom0 completely lost its interrupts (tsc_deadline
was never reset).

After some investigation, I realised the corruption of dom0 interrupt
sequence happened as early as the first external interrupt received on
dom1. Dom1 is configured to ack interrupts and exit. The ack is
necessary to receive the interrupt vector dump that allows dom0 to
understand what caused the exit and call the proper handler.
Unfortunately, the ack moves the IRR bit into the ISR, and subsequent
EOI writes by dom0 seem to be ignored an not reset the APIC correctly.
The fix consists in doing the EOI from tyche (which apparently the host
kvm understands better than from dom0).
Note that cflush on APIC PROC PRI from dom0 was recovering the APIC in
dom0, but any transition to tyche would break interrupts again...

Included in the list of things I had to look at are:
1. PAT for tyche and in dom0. I set the default PAT for tyche and added
   that to the init of dom0.
2. MTRRs and memory types (related to PAT). I think we're not handling
   memory types correctly and the APIC entries might end up cached when
   they shouldn't.
3. APIC was not correctly mapped in first-stage. The code was adding the
   virt offset twice to the virtual address and once to the physical
   one... I fixed it, but this seems to break SMP somehow. TODO: look
   into it.

Current status:
Dom1 boots all the way up to some double fault that I need to
investigate, but survives the htl and kvm_vcpu block. I don't see any
issue with the timer anymore.

SMP is broken on my machine -> figure that out.

Enclaves are probably broken and need some fixes.

Caches were added to both the driver and tyche. We should unify GP
registers and vmcsfields accesses somehow, i.e., do not distinguish
between cache accesses and vcpu.get/vcpu.set when the current context is
set. This will require some refactoring.

Dom1 linux is a bit slow and complains about the drifting of the clock.
See what we can do there, (maybe play with TSC multipliers and the
likes).