tyche: push TlbShootdown to the initiating core
I forgot to update the EPT root on the source core initiating the EPT
root update. It would break the tyche code if I directly update the vcpu
ept root on the domain update site, since we are not supposed to access
the vcpu state when applying the domain updates. An alternative now is
to push the ept root update to all cores, including the source core.
The source core would proceed to free the ept root before updating the
ept root -- however, we will apply the per core updates and process the
TlbShootdown there and update the ept root of the vcpu to the new one
before we enter the vmx non-root mode again.

This is a bit of a hack: we will test and see if it works

Signed-off-by: YuChen Qian <yuchenqian96@gmail.com>