Support for VMs merge with remapper + new sdk
Tyche now supports enclaves and VMs thanks to the new kvm-backend
themis.

It also offers a new sdk with two alternative backends:
1. skd tyche: directly interacts with the tyche driver to create
   domains. DOES NOT SUPPORT INTERRUPT ROUTING.
2. sdk kvm: leverages kvm to create tds. SUPPORTS interrupts
   transparently.

I am gonna clean a lot of things, there are too many back and forth with
adding value and resetting afterwards etc.
Let's just clean that all up.

2. Changing the runtime sdk

We do not need the frame_t anymore, we look for it directly by talking
to tyche. Why? The capa index can change if we had interrupts and were
preempted. This is also a proof of concept of integrating the capa lib
inside and enclave and enumerating its own resource which is a first
step towards checking your are in a confidential environment or spawning
a "nested" domain.