Capa-engine: add support for region revoke capa
The region revoke capabilities are a new kind of capabilities that can
be used to easily revoke regions capabilities (for the new regions).
Until now, the only way to revoke region capa was to own the
capability, which of course does not enable the manager domain to
reclaim regions. With this commit it is now possible to create a revoke
capa from a region capa, thus a manager can create the revoke capa prior
to sending the regions.