tyche/x86: remove INVALID_OPCODE from vmcs exception bitmap
Somehow when we debugged tyche back then, we configured the vmcs in
such a way so that INVALID_OPCODE (exception #6) would cause a vmexit
back to tyche...

Now when I try to log onto the guest in baremetal, an invalid opcode
exception causes a vmexit and tyche crashes.

Remove this bit so that the TD0 guest handles all any software
exceptions.

Signed-off-by: YuChen Qian <yuchenqian96@gmail.com>