Removing old region capas
1. The buffer implementation was bogus
It implemented a stack. Turns out we need a queue.
Assume we create a domain. Then it queues a create domain update.
Assume then we add a region, there is a permission update.
With a stack we do the permission update first while the domain does not
exist...

I implemented a circular buffer queue instead.
I also added some tests

2. Fixed for the access right bug

3. Renaming C(onfidential) to U(nique) + adding alias tests

4. Adding P to capainfo if capa has children (P is for parent)

 5. Fixing bugs + adding tests
The engine was not reducing access rights when a carve happened to
reduce the memops.
The engine was not tracking the regions correclty for a send: sending
a region that had aliased children produced a bug that would give it
access to the entire region.
Also added tests for enclaves doing operations on their own, making sure
the revocation tree works.