Removing tombstone
The tombstone are not needed anymore.
I handle the case of aliasing the same region twice in the driver now.
This allows a simpler and cleaner implementation on the monitor side.
When a remap is active, we expect the client library to undo the remap
as part of the revocation.
This is done with the new revoke_aliased_region monitor call.

On riscv, this does not change anything, we simply ignore the other
arguments and proceed with a normal revoke.