:rocket: The correct config
I had an issue with booting linux and enabling x2apic.
The CONFIG_HYPERVISOR_GUEST seem to be the root cause, but disabling it
led to driver crashes at boot time for td0 (missing msi.domain).
Apparently there is some kind of ordering issue in the linux config and
@charlycst supplied a config whose order apparently just works fine.

This enabled to boot td0 and apparently support the IPIs using x2apic.

As a result, to run the kvm enclaves correctly and td1, I had to replace
the old MMIO-based eoi with the new x2apic one.