Make Tyche boot on hardware with Themis driver
On real hardware KVM was trying to use some AMD MSRs, this commit simply
ignores all AMD MSR reads.
This makes it possible for us to build the Themis driver directly into
our kerne, that way we don't have to wory about module versions when
trying out a new kernel on bare metal.

Finaly, for debugging purpose the kernel printk is redirected to the
serial port on base metal.