Expose attestation serialization through Tyche calls
This commit puts in place the basic infrastructure for serializing
attestations through Tyche calls. This requires getting access to a
user-controlled buffer, which Tyche checks for safety (it must be
accessible by the domain).
For now Tyche is not yet able to write to that buffer, because stage 1
does not map the memory as writable for Tyche.