Forward serialized attestation to trust domain
This commit makes it possible for Tyche to write the result of the
attestation back into the trust domain accessible memory.
Unfortunately, this requires to make all physical memory mapped and
writable. The alternative would be to use register-based communication,
but the attestation can grow quite big and that would need maintaining
state in Tyche.