Add suport for VITAL and CLEANUP bits on regions
The CLEANUP bit guarantees that Tyche will zero-out the memory when the
corresponding region gets revoked (if the region had write access on the
memory).
The VITAL bit kills the domain when the corresponding region gets
revoked, that is useful to guarantee that a test segment is always
present, for instance.