Add feature to remap core IDs on bare metal
Because our bare metal platform has hyperthreading, Linux uses a
different numbering (e.g. first hyper-thread of each core are number
0,1,...7, while second hyper threads are 8,9,...,15, but Tyche sees the
two threads of the first core as core 0 and 1).
This commit adds a remap function that changes the core ID received from
Linux when running on bare metal, which fixes the issue.