Add option to hash a region when sending it
This commit adds support in Tyche for hashing regions when sending them.
At the time of writting the Linux drivers are not yet updated to support
the feature.

This change makes it possible to hash the content of an enclave, which
would bring it closer to feature parity with the old single-enclave
attestation.