MMU-related fixes for the VF2 board. Enclaves work.
Implemented the SV39 paging mode in RISC-V as the VF2 board only supports this one, and previously we were using SV48. Our code is now compatible with both SV48 and SV39. Also optimized PMP-usage. Not keeping a frozen PMP entry anymore - essentially the ACLINT is not protected. Need to configure different PMP entries for the two different platforms - QEMU and VF2. We can keep a frozen entry for protecting ACLINT in QEMU. Also fixed a bug in the PMP crate on identifying if a memory region can be encoded with the NAPOT rule. Added stack alignment for trust domains.