Adding hardware communication benchmark
We want to measure the cost of calling and returning to the monitor.
I added the hwcomm benchmark (env variable HWCOMM).
It runs without workload.
There is a default benchmark setting for it in the config/Makefile.
I tried to add the riscv code as well but @Neelu should double check.

It requires a new monitor call value, which I added to Linux.
Don't forget to update the linux repo as well.