Unified platform-independent monitor implem
New implementation of the monitor logic in a platform independent
manner.
There are new traits for platform state and context, as well as
a monitor trait that takes the platform's state as a parametric type.
The monitor trait implements most of the logic that drives the
capability engine and removes code duplication between riscv and x86.

The x86 implementation further uses the capa engine RegisterState that
includes a caching layer.

Next: do the same for riscv