Fixing stackoverflow in Redis
Redis was crashing when running `command` due to an iovec stack
overflow that was corrupting the BSS segment.
The code in redis allocates a massive buffer on the stack (more than
a page size) to host the iovec when sending a large reply.
This created a stackoverflow (as we only had two pages for the stack),
leading to memory inside the BSS being overwritten. I manifested as
a panic with magic values (secret) being overwritten in the malloc ctx
in musl.

The fix is to move the stack away from the BSS (thus hopefully getting
a page fault for an overflow rather than corrupting memory) and
alocating a larger stack. The fix is self contained within the manifest
of the application and did not require modifying the monitor.