Bugfix: active VMCS can't be migrated
The VMX spec says that active VMCS should not be migrated across cores,
or else there is a risk of VMCS corruption which we were observing on
the machine.
This commit fixes the corruption by issuing a VMCLEAR when switching
frame. This might incur an extra cost which might be worth measuring if
we need to track for inefficiency, but hopefully it doesn't happen too
much on the critical path.